Pass $parentDirectoryId when checking path existence
This fixes not being able to save a file to a path that already exists when the path is more than one directory deep.
For example, with the following folder structure:
```
foo
  | bar
    | foobar.txt
```
`$disk->put("/foo/bar/baz.txt", fopen("baz.txt"))` resulted in this method calling `->requestDirectoryMetadata("bar")`. Because "/bar" doesn't exist on the root directory, it would then try to create "/foo/bar" and throw an exception because "/foo/bar" already exists.